services/api/pyproject.toml
[project]
name = "lifelog-api"
version = "0.1.0"
description = "Lifelog MVP API service"
authors = [{ name = "Lifelog" }]
requires-python = ">=3.11"
dependencies = [
  "fastapi~=0.112",
  "uvicorn[standard]~=0.30",
  "pydantic~=2.7",
  "pydantic-settings~=2.4",
  "PyJWT[crypto]~=2.9",
  "boto3~=1.35",
  "redis~=5.0",
  "celery[redis]~=5.3",
  "qdrant-client~=1.7",
  "httpx~=0.27",
  "prometheus-client~=0.20",
  "python-multipart~=0.0.9",
  "jinja2~=3.1",
  "alembic~=1.13",
  "orjson~=3.10",
  "loguru~=0.7",
  "supabase~=2.4",
  "google-genai>=1.56.0,<2.0.0",
  "google-adk~=1.23",
  "Pillow~=10.4",
  "pillow-heif~=0.16",
  "sqlalchemy[asyncio]~=2.0",
  "asyncpg~=0.29"
]

[project.optional-dependencies]
dev = ["pytest~=8.2", "pytest-asyncio~=0.23", "pytest-cov~=5.0", "ruff~=0.5", "mypy~=1.10"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]

[tool.coverage.run]
source = ["app"]
omit = ["app/db/migrations/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]
